+++ app/src/androidTest/java/woowacourse/payments/ui/cardform/CardFormScreenTest.kt
@@ -1,6 +1,6 @@
 @file:Suppress("ktlint:standard:function-naming")
 
-package woowacourse.payments.ui.newcard
+package woowacourse.payments.ui.cardform
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
@@ -8,9 +8,9 @@ import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import org.junit.Rule
 import org.junit.Test
-import woowacourse.payments.ui.newcard.model.ActionType
+import woowacourse.payments.ui.cardform.model.ActionType
 
-class NewCardScreenTest {
+class CardFormScreenTest {
     @get:Rule
     val composeRule = createComposeRule()
