+++ app/src/main/java/woowacourse/payments/ui/cardform/components/BankSelector.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.newcard.components
+package woowacourse.payments.ui.cardform.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
