+++ app/src/main/java/woowacourse/payments/ui/cardform/components/ExpiryTextField.kt
@@ -1,6 +1,5 @@
-package woowacourse.payments.ui.newcard.components
+package woowacourse.payments.ui.cardform.components
 
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
@@ -12,14 +11,18 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import woowacourse.payments.R
 import woowacourse.payments.designsystem.theme.AndroidpaymentsTheme
-import woowacourse.payments.ui.newcard.model.CARD_HOLDER_MAX
+import woowacourse.payments.ui.common.transformation.GroupedVisualTransformation
+import woowacourse.payments.ui.cardform.model.EXPIRY_MAX
+
+private const val EXPIRY_GROUP_SIZE: Int = 2
+private const val EXPIRY_SEPARATOR: String = " / "
 
 @Composable
-fun CardHolderTextField(
+fun ExpiryTextField(
     value: String,
     onValueChange: (String) -> Unit,
     modifier: Modifier = Modifier,
@@ -29,36 +32,35 @@ fun CardHolderTextField(
     OutlinedTextField(
         value = value,
         onValueChange = { input ->
-            val onlyLettersAndSpace = input.filter { it.isLetterOrDigit() }.take(CARD_HOLDER_MAX)
-            onValueChange(onlyLettersAndSpace)
+            val onlyDigits = input.filter { it.isDigit() }.take(EXPIRY_MAX)
+            onValueChange(onlyDigits)
         },
         modifier = modifier,
-        label = { Text(stringResource(id = R.string.new_card_holder_name_label)) },
-        placeholder = { Text(stringResource(id = R.string.new_card_holder_name_hint)) },
+        label = { Text(stringResource(id = R.string.new_card_expiry_label)) },
+        placeholder = { Text(stringResource(id = R.string.new_card_expiry_hint)) },
         keyboardOptions =
             KeyboardOptions(
-                keyboardType = KeyboardType.Text,
+                keyboardType = KeyboardType.NumberPassword,
                 imeAction = ImeAction.Next,
             ),
         keyboardActions = KeyboardActions(onNext = { onImeAction() }),
-        supportingText = {
-            Text(
-                text = "${value.length}/$CARD_HOLDER_MAX",
-                textAlign = TextAlign.End,
-                modifier = Modifier.fillMaxWidth(),
-            )
-        },
-        singleLine = true,
+        visualTransformation = expiryVisualTransformation,
         colors = colors,
     )
 }
 
+private val expiryVisualTransformation: VisualTransformation =
+    GroupedVisualTransformation(
+        groupSize = EXPIRY_GROUP_SIZE,
+        separator = EXPIRY_SEPARATOR,
+    )
+
 @Preview(showBackground = true)
 @Composable
-private fun CardHolderTextFieldPreview() {
+private fun ExpiryTextFieldPreview() {
     AndroidpaymentsTheme {
-        CardHolderTextField(
-            value = "",
+        ExpiryTextField(
+            value = "1225",
             onValueChange = {},
             modifier = Modifier,
         )
