+++ app/src/main/java/woowacourse/payments/ui/cardform/components/PinTextField.kt
@@ -1,5 +1,6 @@
-package woowacourse.payments.ui.newcard.components
+package woowacourse.payments.ui.cardform.components
 
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
@@ -11,18 +12,14 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import woowacourse.payments.R
 import woowacourse.payments.designsystem.theme.AndroidpaymentsTheme
-import woowacourse.payments.ui.common.transformation.GroupedVisualTransformation
-import woowacourse.payments.ui.newcard.model.EXPIRY_MAX
-
-private const val EXPIRY_GROUP_SIZE: Int = 2
-private const val EXPIRY_SEPARATOR: String = " / "
+import woowacourse.payments.ui.cardform.model.PIN_MAX
 
 @Composable
-fun ExpiryTextField(
+fun PinTextField(
     value: String,
     onValueChange: (String) -> Unit,
     modifier: Modifier = Modifier,
@@ -32,35 +29,30 @@ fun ExpiryTextField(
     OutlinedTextField(
         value = value,
         onValueChange = { input ->
-            val onlyDigits = input.filter { it.isDigit() }.take(EXPIRY_MAX)
+            val onlyDigits = input.filter { it.isDigit() }.take(PIN_MAX)
             onValueChange(onlyDigits)
         },
-        modifier = modifier,
-        label = { Text(stringResource(id = R.string.new_card_expiry_label)) },
-        placeholder = { Text(stringResource(id = R.string.new_card_expiry_hint)) },
+        modifier = modifier.fillMaxWidth(),
+        label = { Text(stringResource(id = R.string.new_card_pin_label)) },
+        placeholder = { Text(stringResource(id = R.string.new_card_pin_hint)) },
         keyboardOptions =
             KeyboardOptions(
                 keyboardType = KeyboardType.NumberPassword,
-                imeAction = ImeAction.Next,
+                imeAction = ImeAction.Done,
             ),
-        keyboardActions = KeyboardActions(onNext = { onImeAction() }),
-        visualTransformation = expiryVisualTransformation,
+        keyboardActions = KeyboardActions(onDone = { onImeAction() }),
+        visualTransformation = PasswordVisualTransformation(),
+        singleLine = true,
         colors = colors,
     )
 }
 
-private val expiryVisualTransformation: VisualTransformation =
-    GroupedVisualTransformation(
-        groupSize = EXPIRY_GROUP_SIZE,
-        separator = EXPIRY_SEPARATOR,
-    )
-
 @Preview(showBackground = true)
 @Composable
-private fun ExpiryTextFieldPreview() {
+private fun PinTextFieldPreview() {
     AndroidpaymentsTheme {
-        ExpiryTextField(
-            value = "1225",
+        PinTextField(
+            value = "",
             onValueChange = {},
             modifier = Modifier,
         )
