+++ app/src/main/java/woowacourse/payments/ui/cardform/model/ActionType.kt
@@ -1,0 +1,6 @@
+package woowacourse.payments.ui.cardform.model
+
+enum class ActionType {
+    NEW,
+    EDIT,
+}
