--- app/src/main/java/woowacourse/payments/ui/newcard/model/ActionType.kt
@@ -1,6 +1,0 @@
-package woowacourse.payments.ui.newcard.model
-
-enum class ActionType {
-    NEW,
-    EDIT,
-}
